Documented ID service and reorganized logging service
