+++ lib/src/id/id_enums.dart
@@ -1,0 +1,14 @@
+part of "id_service.dart";
+
+/// Enum representing different types of IDs managed by the `ArcaneIdService`.
+///
+/// The `ID` enum has two possible values:
+/// - `session`: Represents the session ID.
+/// - `install`: Represents the install ID.
+enum ID {
+  /// Represents the session ID.
+  session,
+
+  /// Represents the install ID.
+  install,
+}
