+++ lib/src/logging/logging_service.dart
@@ -7,6 +7,8 @@ import "package:flutter/foundation.dart";
 
 export "package:logger/logger.dart" show Level;
 
+part "logging_interface.dart";
+
 /// A singleton class that manages logging to one or more logging interfaces
 /// with optional metadata.
 ///
@@ -225,9 +227,6 @@ class ArcaneLogger {
 
     for (final LoggingInterface i in interfaces) {
       I._interfaces.add(i);
-      if (i is ArcaneDebugConsole) {
-        await I._interfaces.firstWhere((x) => x == i).init();
-      }
     }
 
     return I;
@@ -323,43 +322,3 @@ class ArcaneLogger {
   /// Clears all persistent metadata.
   void clearPersistentMetadata() => _additionalMetadata.clear();
 }
-
-/// Represents a logging interface that can log messages to different destinations.
-///
-/// Concrete implementations of this class should override the [log] method to provide
-/// platform-specific logging behavior.
-abstract class LoggingInterface {
-  LoggingInterface._internal();
-  static late final LoggingInterface _instance;
-
-  /// Provides access to the singleton instance of the `LoggingInterface`. This
-  /// ensures that the logging interface, once configured, remains so.
-  static LoggingInterface get I => _instance;
-
-  final bool _initialized = false;
-
-  /// Whether the logging interface has been initialized.
-  bool get initialized => I._initialized;
-
-  /// Initializes the logging interface.
-  ///
-  /// If any configuration needs to be performed on the logging interface prior
-  /// to use, this is where it should be done.
-  /// This method should, at a minimum, set `I._initialized = true`.
-  Future<LoggingInterface?> init();
-
-  /// This method is called by the `ArcaneLogger` when a log message is
-  /// received. See `ArcaneLogger.log` for further details on how logging
-  /// works and what options are available.
-  void log(
-    String message, {
-    Map<String, dynamic>? metadata,
-    Level? level,
-    StackTrace? stackTrace,
-  });
-}
-
-/// A special logging interface which specifically targets the local debug
-/// console. This `LoggingInterface` differs in that it is immediately
-/// initialized when it is registered with the `ArcaneLogger`.
-abstract class ArcaneDebugConsole implements LoggingInterface {}
